pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apg"
version = "0.1.0"
description = "Agent Policy Gateway - Cedar-based intent execution control for LLMs and agents"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Travis Langley"},
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pyyaml>=6.0",
    "pyjwt[crypto]>=2.8",
    "cedarpy>=0.3",
    "click>=8.1",
    "pydantic>=2.0",
    "watchfiles>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
apg = "apg.cli:cli"

[tool.setuptools.packages.find]
include = ["apg*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=apg --cov-report=term-missing"
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
